Cleanup Veracode integration.
These are mostly style fixes: making this a bit more PEP8 compliant,
using defaultdict to simplify some code, using python idioms for things
like checking if a list is empty, etc.

The unit test still passes, but it's a bit brittle. It fails when run
against Melbourne. We might want to better document what needs to
happen for it to work.