Preliminary support for Fortify import - Cleanup needed
This works similar to the Veracode import since we use the base class
for matching issues with tasks.

The Fortify mapping file is generated by an XSLT - run against a Fortify
rules file (with Internal and External category mapping). The present
implementation re-uses the same code. For instance the fortify mapping
file uses a category name; whereas the expectation was that it would be
a CWE-ID.

One important note is that the report id is not extracted from the
Fortify report. Since SD Elements does not presently have support for
'fortify' we use 'veracode' for the time being.